+++ cmd/kview/main.go
@@ -105,6 +105,7 @@ func main() {
 	srv.Actions().Register("namespaces.delete", kubeactions.HandleNamespaceDelete)
 
 	srv.Actions().Register("customresourcedefinitions.delete", kubeactions.HandleCRDDelete)
+	srv.Actions().Register("customresource.delete", kubeactions.HandleCustomResourceDelete)
 	srv.Actions().Register("custom.workload", kubeactions.HandleCustomWorkloadAction)
 	srv.Actions().Register("resource.yaml.validate", kubeactions.HandleResourceYAMLValidate)
 	srv.Actions().Register("resource.yaml.apply", kubeactions.HandleResourceYAMLApply)